Move graceful_exit to util.py
